+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2024-01-04
+
+### Fixed
+
+ - Added plugin schema to fix serverless warning. Thank you @crunchie84 ([74](https://github.com/amplify-education/serverless-log-forwarding/pull/74))
+ - Refactored integration tests
+ - Updated linting
+ - Updated packages
+
 ## [3.1.0] - 2023-02-17
 
 ### Added